[Refactor](json-schema): move getResolveSchema helper
